+++ backend/.pre-commit-config.yaml
@@ -1,5 +1,5 @@
 default_stages: [commit]
-fail_fast: true
+fail_fast: false
 repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v4.0.1
